Fix inaccurate comment
The previous commit was merged somewhat prematurely, and fixed the
coverage on Travis by limiting `kcov` to a single process.

The previous commit also pins to a specific branch, which was not
necessary to fix this issue, but is a good idea anyway to prevent
bad commits on kcov's master branch from breaking PyO3's build.