Add mocpy to the list of examples
MOCpy is an astronomical Python library used for describing arbitrary
coverage regions on the sky unit sphere. It relies on the HEALPix tesselation
scheme and because of this, it is very easy and fast to compute the
intersection/union... between coverage regions!

Please check the documentation of mocpy [here](https://cds-astro.github.io/mocpy/) if you are curious.