add experimental support for PEP 489 multi-phase module initialization
This commit adds experimental but *almost* fully functional support
for multi-phase module initialization as specified in PEP 489.

With the exception of adding submodules to modules, all other
functionality should be retained. In other words, unless submodules
are used, multi-phase initialization is a complete drop-in
replacement.

Note that this commit serves as a demonstration & basis for further
improvements only; many tests have therefore not been adapted
correspondingly yet.

Nevertheless, it runs.

Signed-off-by: Max Carrara <m.carrara@proxmox.com>