[RFC] split PyErr::new() into multiple methods
Introduces new_empty(), new_arg() and new_args(), and
the same on Exception types as new_err_empty(), new_err_arg() and
new_err_args().

After the deprecation cycle, new_arg() could be renamed back to new(),
since it is the one used in the vast majority of cases.

I kept `PyErrArguments`, but it's now only implemented for tuples,
not for all Python-convertible types.