[RFC] split PyErr::new() into multiple methods
Similar to call(), introduces new0(), new1() and new_args(), and
the same on Exception types as new_err0(), new_err1() and
new_err_args().

After the deprecation cycle, new1() could be renamed back to new(),
since it is the one used in the vast majority of cases.

I kept `PyErrArguments`, but it's now only implemented for tuples,
not for all Python-convertible types.