Fix ambiguous associated item error (#4725)
* Add test for ambiguous associated item

The `#[pyclass]` macro implements `IntoPyObject` for the annotated enum.
When the enum has any of `Error`, `Output` or `Target` as members, this
clashes with the associated types of `IntoPyObject`.

This also happens when deriving `IntoPyObject` directly.

* Fix #4723: ambiguous associated item in #[pyclass]

This uses the fix described in https://github.com/rust-lang/rust/issues/57644

Also apply fix to `#[derive(IntoPyObject)]`.