WIP: Generalize the rpath API for relocatable Python installations
Relocatable Python installations like python-build-standalone have a
similar need for an rpath as the macOS framework builds in Xcode.
Furthermore, the platform on which you build your code is not really a
function of the code itself. This change generalizes the recent
`add_python_framework_link_args()` from #4833 to
`add_python_link_args()`, which also does the right thing on
python-build-standalone, and encourages everyone to set up a build.rs
file that invokes it, whether or not they need it on their current
machine.