Fix `Probe` class naming issue with `#[pymethods]` (#4988)
Commit 603a55f204e3 ("fix `#[pyclass]` could not be named `Probe`
(#4794)") fixed the issue where it was not possible to define a class
named `Probe`. However, similar issue exists when trying to add
`#[pymethods]` implementations to the `Probe` class. It generates
similar confusing errors as the original issue (#4792), due to the
internal `Probe` trait being in scope at the user code:

  error[E0782]: expected a type, found a trait
  help: you can add the `dyn` keyword if you want a trait object
  # | impl dyn Probe {

Fix the issue by avoiding pollution of the imports from the macro
expansion and instead use qualified path to the internal types.