Unraw complex struct enum field python names
Stray case related to #2395 (probably unnoticed because the functionality is
newer and a bit of an edge case).

The `python_name` of struct-type enum variant fields (and the `__match_args__`
of the variant) do not unraw the names of these fields, which effectively
makes them unusable as the resulting identifiers contain '#' which the Python
lexer interprets as comments. The only reason you'd probably want to do this
anyway is to use reserved identifiers (`type` in my case).