src/impl_: add pymodule_state for core ModuleState
This commit adds the core ModuleState & StateCapsule structs for
interop with python's ModuleDef m_size field.

We choose to keep the in-interpreter size small, only being a NonNull
ptr to the actual (allocated by Rust) state. This is suboptimal from
a pointer chasing perspective, because we'll eventually have
interpreter->StateCapsule->TypeMap<T> dereference chains, but it makes
it easier to implement and reason about.